+++ src/components/elements/StyledSelect.tsx
@@ -14,15 +14,19 @@ export default function StyledSelect({
     name,
     label,
     options,
-}: StyledSelectArgs) {
+    compact
+}: Readonly<StyledSelectArgs>) {
     const { t } = useTranslation();
 
     return (
-        <div className={label !== false ? "relative z-0 w-full group" : "relative z-0 w-full group"}>
+        <div className={compact ? "w-full group flex flex-row items-center justify-between" : "relative z-0 w-full group"}>
             {label !== false && (
                 <label
                     htmlFor={name}
-                    className="block mb-1 text-sm font-medium text-left text-gray-700 dark:text-gray-300"
+                    className={
+                    compact ? "block mr-1 text-sm font-medium text-left text-gray-700 dark:text-gray-300" :
+                        "block mb-1 text-sm font-medium text-left text-gray-700 dark:text-gray-300"
+                    }
                 >
                     {label || name}
                 </label>
@@ -32,9 +36,11 @@ export default function StyledSelect({
                 id={name}
                 value={value}
                 onChange={onSelect}
-                className="block w-full px-3 py-2 text-sm text-gray-900 bg-white border border-gray-300 rounded-md shadow-sm
-                  focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700
-                  dark:text-white dark:border-gray-600 dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                className={
+                    (compact ? "px-2 py-1.5 w-fit" : "px-3 py-2 w-full") +
+                    " block text-sm text-gray-900 bg-white border border-gray-300 rounded-md shadow-sm" +
+                    " focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700" +
+                    " dark:text-white dark:border-gray-600 dark:focus:ring-blue-500 dark:focus:border-blue-500"}
                 required
             >
                 <option defaultChecked={true}>{t('Please Select')}</option>